reorganize python files
