+++ inst/python/finalize_spacyPython.py
@@ -1,9 +1,9 @@
-#!/usr/local/bin/python
-
-import gc
+# this script remove spobj and also spacyNLP to free up memory
 
 if "spobj" in locals():
     del spobj
 
-del nlp
+if "nlp" in locals():
+    del nlp
+
 gc.collect()
